fix(dtype): address PR #72 bot review -- 4 missed upcasts + regression tests
Gemini and Copilot both caught real regressions I introduced in the
earlier commit.  Ironic given the stated goal was to eliminate silent
upcasts, exactly validates the reason bots review PRs.

Gemini found four sites where a Python-literal or Python-float
divisor was upcasting a float32 array to float64 under pre-NEP-50
numpy (>= 1.24 in pyproject), now not tapped back by any astype:

- _ivfpq.py:257  rot /= einsum + 1e-12
- _ivfpq.py:271  q / q_norm  (q_norm = float(np.linalg.norm(q)))
- _ivfpq.py:983  q_pre_all /= einsum + 1e-12  (search_batch path)
- _pq.py:190     rot /= einsum + 1e-12
- _pq.py:207     q / q_norm
- _residual.py:229  q / q_norm
- _index.py:427  q / q_norm

Fix the root cause by wrapping each scalar in np.float32():
  np.float32(1e-12), np.float32(q_norm), np.float32(np.linalg.norm(rot))

Copilot suggested regression tests for these dtype invariants.
Added tests/test_dtype_invariants.py with 5 checks:

- rht() returns float32 for float32 input (1-D and 2-D, across all
  supported pdim values)
- SnapIndex.add_batch stores float32 _norms
- PQSnapIndex._preprocess + _preprocess_single return float32 under
  all three branches (plain, use_rht, use_opq)
- IVFPQSnapIndex._preprocess + _preprocess_single + the separate
  search_batch normalisation path all stay float32 under the same
  three branches
- ResidualSnapIndex search path keeps stored _norms float32

All 204 tests pass (199 existing + 5 new), mypy --strict clean, ruff
clean.  Two cast() additions needed on returns of 'q / np.float32(q_norm)'
because numpy stubs do not infer the scalar wrap correctly.

+++ snapvec/_residual.py
@@ -226,7 +226,9 @@ class ResidualSnapIndex(FreezableIndex):
         if q_norm < 1e-10:
             return []
         pdim = self._pdim
-        q_unit = q / q_norm
+        # Python float ``q_norm`` is float64 under pre-NEP-50 numpy; wrap
+        # so ``q_unit`` stays in float32.
+        q_unit = q / np.float32(q_norm)
         q_padded = np.zeros(pdim, dtype=np.float32)
         q_padded[: self.dim] = q_unit
         q_rot = rht(q_padded[None, :], self.seed)[0]
